Renammed file sanitize_xml_workspace
