+++ modules/blockly/utils/sanitize_xml_workspace.js
